Python-SDK-implementation-for-RightSizing-Rules-CRUD,-attach-and-detach (#166)

+++ spotinst_sdk2/models/ocean/aws/__init__.py
@@ -965,7 +965,7 @@ class Roll:
 class AutoScaleDown:
     """
     # Arguments
-    max_scale_down_percentage: double
+    max_scale_down_percentage: float
     """
 
     def __init__(
